requirements.txt
flask
tensorflow
tensorflow_hub
Pillow
numpy
